requirements.txt
# -*- coding: UTF-8 -*-
yfinance
schedule
requests
python-dotenv
